Adding option "CaseClassSkipNulls"
Usually the annotation @JsonInclude(Include.NON_NULL) will cause null fields not to be emitted. However,
in the case of serialization of an Option[] field containing None, the standard JacksMapper will still write "null":
since the field itself is not null, even when the flag is used None is serialized as null, regardless.

This behavior cannot be changed because of compatibility reasons. Hence, this option "CaseClassSkipNulls" is introduced,
with behavior complementary to CaseClassCheckNulls. When this option is selected, each Option field that contains None will
be skipped during serialization, rather than being serialized as null. This option does not override "ALWAYS" or other
Jackson flags explicitly set on classes or individual fields: if you do, the flags will still be honored.