Deserialize missing Option case class fields as None
The version 2.1.4 of Jacks deserializes a missing property corresponding
to an Option case class field as "null", rather than None. This leads to
situations like:

Start with ""
Deserialize -> Z(null)
Serialize -> """{"s":null}"""
Deserialize -> Z(None)

Which may be undesirable. When the option CaseClassCheckNulls is in use,
nulls should never be generated; therefore, deserialize missing properties
that correspond to Option fields as None instead, unless an explicit
default is specified.